Move HTTP update to separate file
